requirements.txt
aiohttp
aioodbc
pydantic
python-dotenv
sqlalchemy